+++ src/main/java/lotto/global/Error.java
@@ -5,7 +5,9 @@ public enum Error {
     DUPLICATED("로또 번호는 서로 중복될 수 없습니다."),
     INVALID_TYPE("숫자만 입력 가능합니다."),
     INVALID_UNIT("1,000원 단위로만 입력 가능합니다."),
-    INVALID_AMOUNT_RANGE("구매금액은 1,000원 이상 100,000원 이하여야 합니다.");
+    INVALID_AMOUNT_RANGE("구매금액은 1,000원 이상 100,000원 이하여야 합니다."),
+    INVALID_NUMBER_RANGE("로또 번호는 1-45 사이의 숫자만 입력 가능합니다."),
+    INVALID_BONUS_NUMBER("보너스 번호는 로또번호와 중복될 수 없습니다.");
     private static final String ERROR = "[ERROR] ";
     private final String message;
